test: Move acceptance test fixtures to conftest
Remove scope from populated_database. Introspect request to call
module-specific setup_data.
